backend/requirements.txt
Flask
Pillow
Werkzeug
